03-Introduccion-ORM/requirements.txt
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1
python-dotenv==1.0.0


# Core ORM and Database
SQLAlchemy==2.0.23
alembic==1.12.1

# Database Drivers
psycopg2-binary==2.9.9  # PostgreSQL
PyMySQL==1.1.0          # MySQL
sqlite3                 # SQLite (built-in with Python)

# Data Validation and Serialization
pydantic==2.5.0
email-validator==2.1.0

# Additional Utilities
python-dotenv==1.0.0    # Environment variables
typing-extensions==4.8.0  # Type hints support
